backend/requirements.txt
flask
flask-cors
python-dotenv
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-subscription
prometheus-api-client==0.5.3
tinydb
